last modification after eclipse error

+++ WS-Project/src/main/java/com/ws/controller/UsersController.java
@@ -25,6 +25,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -51,48 +52,16 @@ public class UsersController {
 	@Autowired
 	GroupeService grService;
 	
-    @Autowired
-	private ConnectionRepository connectionRepository;
-	private Twitter twitter;
+	    
+	   
+	    
+	    
+	   
 	
 	    
 	    
-	    
-	    @Inject
-	    public UsersController(Twitter twitter, ConnectionRepository connectionRepository) {
-	        this.twitter = twitter;
-	        this.connectionRepository = connectionRepository;
-	    }
-
-	   
-	    
-	   
-	    
-	    
-	    
-	   
-	    @GetMapping
-	    public  String helloTwitter(Model model, Principal currentUser) {
-	        if (connectionRepository.findPrimaryConnection(Twitter.class) == null) {
-	            return "login";
-	            
-	            
-	        }
 	 
-	        
-	        model.addAttribute("ID" , twitter.userOperations().getUserProfile());
-	        model.addAttribute("profil" , twitter.userOperations().getUserProfile());
-	        CursoredList<TwitterProfile> friends = twitter.friendOperations().getFriends();
-	        model.addAttribute("friends", friends);
-	        User current=service.findByFirstName(twitter.userOperations().getUserProfile().getName());
-	        service.setCurrent(current);
-	        
-	        return "redirect:/user/home";
-	    }
-	    
-	    
-	
-	
+	 
 	 @GetMapping("/user/home")
 	  public String home(HttpServletRequest request ) {
 		 if(service.getCurrent()==null){
@@ -101,13 +70,12 @@ public class UsersController {
 		 }
 		 
 		 
-		 
-
-		 request.setAttribute("current", service.getCurrent());
-		 
-		 request.setAttribute("users", service.findAll());
-		 
-	        return "userList";
+		
+      
+		 User user=service.findOne(service.getCurrent().getEmail());
+		 request.setAttribute("user", user);
+		
+	        return "userHome";
 	 }
 	 
 	 
@@ -127,8 +95,15 @@ public class UsersController {
 	 
 	 
 	    @GetMapping("/user/delete/{id}")
-		public String delete(@PathVariable int id, ModelMap map) {
+		public String delete(@PathVariable String id, ModelMap map) {
+	    
+	    	
+	    if(service.getCurrent().getEmail().equals(id)){
+	    	
+	    	return "redirect:/login/home";
+	    }
 		 
+    	
 		 service.delete(id);
 		 
 		 return "redirect:/user/home";
@@ -165,23 +140,29 @@ public class UsersController {
 	 
 	 
 	 @RequestMapping(value = "/user/edit/{id}", method = RequestMethod.GET)
-	    public String edit(@PathVariable int id,
+	    public String edit(@PathVariable String id,
 	                       Model model) {
 	        User user = service.findOne(id);
+	        
+	       
+	        
+	       
+	        
 	        model.addAttribute("user", user);
 	        return "editUser";
 	    }
 	  
 	 
 	
-	  @RequestMapping(value = "/user/update", method = RequestMethod.POST)
-	    public String update(@RequestParam("id") int id,
+	 // @RequestMapping(value = "/user/update", method = RequestMethod.POST)
+	  @PostMapping(value = "/user/update")
+	    public String update(
 	                               @RequestParam("email") String email,
 	                               @RequestParam("firstName") String firstName,
 	                               @RequestParam("lastName") String lastName,
 	                               @RequestParam("biography") String biography) {
-		  User user = service.findOne(id);
-		  user.setEmail(email);
+		  User user = service.findOne(email);
+		  
 	      user.setFirstName(firstName);
 	      user.setLastName(lastName);
 	      user.setBiography(biography);
@@ -190,259 +171,7 @@ public class UsersController {
 	    }
 	 
 	 
-	  /*
-	   * Gestion group
-	   *  
-	   */
-
-		 @GetMapping("/group/home")
-		 public String  home(ModelMap model){
-			 
-			 if(service.getCurrent()==null){
-				 return "redirect:/";
-				 
-			 }
-			 
-			 
-	        
-			model.addAttribute("groups" , grService.findAll());
-			//grService.findByAdmin(service.getCurrent())
-			
-			model.addAttribute("myOwnGroup" ,service.getCurrent().getMyownGroup());
-			
-			model.addAttribute("myJoinGroup" ,service.getCurrent().getMyJoingroup());
-			 
-			 return "groupeList";
-		 }
-		 
-	 
-	  @GetMapping("/group/new")
-	    public String groupForm(ModelMap map) {
-		  
-		  if(service.getCurrent() == null){
-				 return "redirect:/";
-			 }
-			 
-			 
-	        map.addAttribute("group", new Group());
-	        
-	        return "newGroup";
-	    }
-	 
-	 
-	 @PostMapping("/group/create")
-	 public String CreateGroup(@ModelAttribute(value="group" )Group group ) {
-		    
-		  
-		 
-		 
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		 
-		 
-		 group.setId(grService.findAll().size()+1);
-		 group.setAdmin(service.getCurrent());
-		 grService.save(group);
-	     return "redirect:/group/home";
-	    }
-	 
-	 
-	 
-	 
-	 
-	 @RequestMapping(value = "/group/edit/{id}", method = RequestMethod.GET)
-	    public String editGroup(@PathVariable int id,
-	                       Model model) {
-		 
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		 
-		 Group group = grService.findOne(id);
-	      model.addAttribute("group", group);
-	      return "editGroup";
-		 
-		 
 	
-	    }
-	 
-	
-	  
-	 
-	 @RequestMapping(value = "/group/update", method = RequestMethod.POST)
-	    public String update(@RequestParam("id") int id,
-	                               @RequestParam("name") String name,
-	                               @RequestParam("description") String description
-	                               ) {
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		 
-		 Group group = grService.findOne(id);
-		  group.setId(id);
-		  group.setName(name);
-		  group.setDescription(description);
-		  grService.save(group);
-	      return "redirect:/group/home";
-	    }
-	 
-	 
-	   
-	   @GetMapping("/group/delete/{id}")
-		public String deleteGroup(@PathVariable int id, ModelMap model) {
-		 
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		
-	        grService.delete(id);
-	        
-	        service.getCurrent().delete(id);
-			return "redirect:/group/home";
-			
-		}
-	   
-	   
-	   @GetMapping("/group/leave/{id}")
-		public String leave(@PathVariable int id, ModelMap model) {
-		 
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		
-		   service.getCurrent().leaveGroup(id);
-			return "redirect:/group/home";
-			
-		}
-	   
-	   
-	   
-	   
-	   @GetMapping("/group/join/{id}")
-		public String Join(@PathVariable int id, ModelMap model) {
-		 
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		
-		 Group group = grService.findOne(id);
-		   service.getCurrent().joinGroup(group);
-			return "redirect:/group/home";
-			
-		}
-	   
-	   
-	 
-      
-	   
-	   // addcomment
-
-	 
-	   
-	   @RequestMapping(value = "/group/addComment/{id}", method = RequestMethod.GET)
-	    public String addMember(@PathVariable int id,
-	                       Model model) {
-		 
-		 
-
-		 if(service.getCurrent() == null){
-			 return "redirect:/";
-		 }
-		 
-		
-		 
-		 Group group = grService.findOne(id);
-		 System.out.println(group.getId());
-		 model.addAttribute("group", group);
-		 model.addAttribute("user", service.getCurrent());
-		 model.addAttribute("commentaire", new Comment());
-	      
-	      
-	      return "addComment";
-		 
-		 
-	
-	    }
-	   
-	   @RequestMapping(value = "/group/updateComment", method = RequestMethod.POST)
-	    public String updateComment(@RequestParam("group") int idgroup,
-	    		                   @RequestParam("userId") int userId,
-	                               @RequestParam("comment") String comment) {
-		   
-		   
-		   
-		 
-	      
-	      if(service.getCurrent() == null){
-				 return "redirect:/";
-			 }
-	      
-	       User user=service.findOne(userId);
-		   Comment com=new Comment(user, comment); 
-		  
-		   Group group = grService.findOne(idgroup);
-		   com.setId(group.getComments().size()+1);
-		   group.addComment(com);
-		   grService.save(group);
-		   
-		      return "redirect:/group/home";
-	    }
-
-		 
-	     
-
-		
-	   
-	      
-	   
-	   
-	  
-	   
-	   @RequestMapping(value = "/group/updateMember", method = RequestMethod.POST)
-	    public String updateMember(@RequestParam("group") int idgroup,
-	    		
-	    		                  
-	                               @RequestParam("email") String email,
-	                               @RequestParam("firstName") String firstName,
-	                               @RequestParam("lastName") String lastName,
-	                               @RequestParam("biography") String biography) {
-		   
-		  User user = new User(service.findAll().size()+1, email, firstName, lastName, biography);
-		  
-		 // user.setId(service.findAll().size()+1);
-	      service.save(user);
-	      if(service.getCurrent() == null){
-				 return "redirect:/";
-			 }
-			 
-			 Group group = grService.findOne(idgroup);
-			  
-			  group.addMembers(user);
-			  grService.save(group);
-		      return "redirect:/group/home";
-	    }
-
-
-
-
-
-
-
-
-
-
-
